pyproject.toml
[tool.poetry]
name = "minidream"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9, <4.0"
pre-commit = "^3.6.0"
tinygrad = {git = "https://github.com/tinygrad/tinygrad.git"}
ipykernel = "^6.29.0"
jupyter = "^1.0.0"
torch = "^2.1.2"
torchvision = "^0.16.2"
wandb = "^0.16.2"
einops = "^0.7.0"
gymnasium = "^0.29.1"
furuta = {git = "https://github.com/Armandpl/furuta"}
hydra-core = "^1.3.2"
pygame = "^2.5.2"
torchrl = "^0.3.0"
matplotlib = "^3.8.2"
pytest = "^8.0.1"
minatar = "^1.0.15"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
